"Changes and fixes based on initial review"
